requirements.txt
requests==2.28.2
flask==1.1.4
markupsafe==2.0.1
Pillow==10.1.0